refactor walk to utils/filesystem
